% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vet.R
\name{vet}
\alias{vet}
\title{Calcula o Valor Esperado da Terra}
\usage{
vet(taxa, custo_formacao, preco_madeira, producao, ciclo)
}
\arguments{
\item{taxa}{Numeric. Taxa de desconto na forma decimal}

\item{custo_formacao}{Numeric. Valor presente do custo de formacao}

\item{preco_madeira}{Numeric. Preco de venda da madeira em pé}

\item{producao}{Numeric. Producao de madeira no fim do ciclo}

\item{ciclo}{Numeric. Idade do ciclo}
}
\description{
Convenient function to save charts in Eps format.
}
