% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vet.R
\name{vet}
\alias{vet}
\title{Calcula o Valor Esperado da Terra}
\usage{
vet(taxa, rotacao, vp_custo, producao, preco_madeira)
}
\arguments{
\item{taxa}{Numeric. Taxa de desconto na forma decimal}

\item{rotacao}{Numeric. Idade do ciclo}

\item{vp_custo}{Numeric. Valor presente do custo de formacao}

\item{producao}{Numeric. Producao de madeira no fim do ciclo}

\item{preco_madeira}{Numeric. Preco de venda da madeira em pé}
}
\description{
Convenient function to save charts in Eps format.
}
