% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vet.R
\name{vet}
\alias{vet}
\alias{test_pkgs}
\alias{check_pkgs}
\title{Test an check multiple packages.}
\usage{
vet(pkgs, .f, parent = fs::path_dir(usethis::proj_get()))

test_pkgs(pkgs, parent = fs::path_dir(usethis::proj_get()))

check_pkgs(pkgs, parent = fs::path_dir(usethis::proj_get()))
}
\arguments{
\item{pkgs}{Character string of the names of packages to vet.}

\item{.f}{A symbol; either devtools::test or devtools::check.}

\item{parent}{The absolute path to the parent directory of all packages.
You can omit this if your working directory is one of the packages under
the `parent` directory.}
}
\value{
Invisible.
}
\description{
Vectorized versions of [devtools::test()] and [devtools::check()] designed
to work with multiple packages stored under a single parent directory.
}
\examples{
\dontrun{
pkgs <- c("fgeo.tool", "fgeo.abundance")
test_pkgs(pkgs)
check_pkgs(pkgs)
}
}
\seealso{
[devtools::test()] and [devtools::check()].
}
